Utilise `sysfs` data over `ip a` parsing
Instead of manually parsing the output of `ip a`, we want to use the
built-in ability from `/sys/` to detect state. This is done via the
`operstate` file where we can check if it's connected or not.
Information via [[http://unix.stackexchange.com/a/150323]].

We will only get an `up` status when the device actually has a
connection, not when i.e. `ip l set <device> up` has been run.

Affects: [BSPWM]