Amend our aliases to get Git root directory
Remove the annoying hyphen in `git-root` which moves our hands off the
home row, and makes it less easy to tabcomplete.

Additionally, alias `groot` to the `cd ...` functionality, and have
`gitroot` as a shorthand alias for the `git rev-parse` to make it
quicker, and not reliant on memory.

Affects: [Term]