Display our workspace name more visibly
We want to be able to know which workspace we're in, regardless of
where in the folder hierarchy we are. This is especially useful in very
large directory trees.

For such a tree, it is useful to have a prompt, such as:

	jamie@TheColonel W:bspwm-340 tests/desktop (git)-[issue-340] %

Which indicates we are in the 'bspwm-340' workspace, inside the
'tests/desktop' folder.

With a much larger tree, such as:

	jamie@TheColonel W:lorem-ipsum ..e/consequatur/impedit/dolore/ipsum/aut %

It is useful to be able to shorten the full path to 40 characters, with
a '..' to indicate the truncation.

Affects: [Term]