Migrate to Spotify for music
I've finally moved to the official Spotify application as I've found
that offline support is something I'd really like.

We also need to move to `playerctl` which can interface with Spotify.

We need to create a new `spotify-current.rb` script to be called by
Polybar as it doesn't support it natively.