Use Ruby wrapper for battery indicators
It's getting annoying that the Bash version was having issues comparing
numbers, so to save some complexity, we'll refactor it out into Ruby.

This unfortunately does mean that there's a Ruby dependency here, which
means out-of-the-box it may not always work.

This does also require a somewhat hacky way to interface with the
`.theme_colours.h` file, instead of the generated XResources DB (so we
have to parse that file, and if we change its setup, it'll break).